removed dependencies
