+++ frontend/src/js/controllers/preset-controller.js
@@ -1,7 +1,6 @@
 import * as Config from "../config";
 import * as DB from "../db";
 import * as Notifications from "../elements/notifications";
-import * as Settings from "../pages/settings";
 import * as TestLogic from "../test/test-logic";
 import * as TagController from "./tag-controller";
 
@@ -19,7 +18,6 @@ export function apply(_id) {
       TestLogic.restart();
       Notifications.add("Preset applied", 1, 2);
       Config.saveToLocalStorage();
-      Settings.update();
     }
   });
 }
